Modify API to work with pre-computed embeddings
OpenBB documentation is now represented via pre-computed embeddings,
which makes the startup faster. Also, PDM is included to manage the
project dependencies as pip was too slow when resolving them and the
Dockerfile is adjusted accordingly. The tests on Windows are removed
as the API is meant for Linux servers.