Modify API to work with AutoLlamaIndex
AutoLlamaIndex uses llama-index for a cleaner and faster retrieval and
generation. It provides support for OpenAI models, which are the default
ones now in this API. Furthermore, using its capabilities the API has
two operating modes: programmer and agent. A field in the post request
allows selecting one or the other.