Modify API to work with multiple PATs
Now the code is executed outside the agent, so that the PAT received in
the request can be included in the execution. Previously, the PAT was
overrided before it was executed in the agent, which resulted in always
using the same PAT when multiple requests arrived at the same time. A
test is also added to create the situation where multiple concurrent
requests arrive to the API.